+++ src/main/java/com/infoshare/alpha/wwr/domain/facilities/dao/FacilitiesRepositoryDao.java
@@ -2,7 +2,7 @@ package com.infoshare.alpha.wwr.domain.facilities.dao;
 
 import com.infoshare.alpha.wwr.domain.facilities.entity.Facilities;
 import com.infoshare.alpha.wwr.domain.facilities.entity.Facility;
-import com.infoshare.alpha.wwr.utils.FacilitiesException;
+import com.infoshare.alpha.wwr.utils.exceptions.FacilitiesException;
 
 public interface FacilitiesRepositoryDao {
