+++ src/main/java/com/infoshare/alpha/wwr/domain/patients/dao/PatientsRepositoryDao.java
@@ -1,0 +1,11 @@
+package com.infoshare.alpha.wwr.domain.patients.dao;
+
+import com.infoshare.alpha.wwr.domain.patients.entity.Patients;
+
+public interface PatientsRepositoryDao {
+
+	public void persist(Patients patients);
+
+	public Patients getAll();
+
+}
