--- src/main/java/com/infoshare/alpha/wwr/domain/patients/readmodel/PatientsReadModelDb.java
@@ -1,13 +1,0 @@
-//package com.infoshare.alpha.wwr.domain.patients.dao;
-//
-//import com.infoshare.alpha.wwr.domain.patients.entity.Patients;
-//import com.infoshare.alpha.wwr.domain.patients.query.PatientQuery;
-//
-//public interface PatientsReadModelDb {
-//
-//	public void persist(Patients patients);
-//
-//	public Patients getAll();
-//
-//	public Patients getByQuery(PatientQuery patientQuery);
-//}
