--- src/main/java/com/infoshare/alpha/wwr/domain/patients/repository/PatientsRepository.java
@@ -1,8 +1,0 @@
-//package com.infoshare.alpha.wwr.domain.patients.repository;
-//
-//import com.infoshare.alpha.wwr.domain.patients.entity.Patients;
-//
-//public interface PatientsRepository {
-//
-//    void persist(Patients patients);
-//}
